dtype fix running bench

+++ src/rotorquant_vllm/triton/isoquant_decompress.py
@@ -85,7 +85,7 @@ def iso_decompress(packed: torch.Tensor,
     norms_cont = norms.reshape(M).contiguous().float()
     centroids_f32 = centroids.float().contiguous()
 
-    out = torch.empty(M, D, device=packed.device, dtype=torch.float32)
+    out = torch.empty(M, D, device=packed.device, dtype=dtype)
 
     try:
         _iso_decompress_kernel[(M,)](
